requirements.txt
transformers
datasets
evaluate
torch
tqdm
pyyaml
streamlit
transformers[torch]